Move pages dir into components dir
