+++ src/components/pages/CapturePage/CapturePage.tsx
