+++ src/components/pages/OptionsPage/OptionsPage.tsx
@@ -1,6 +1,6 @@
 import Webcam from "react-webcam";
 import OptionsForm from "./OptionsForm/OptionsForm";
-import { PhotoboothFormProps } from "../../types/photobooth";
+import { PhotoboothFormProps } from "../../../types/photobooth";
 const OptionsPage = ({
   webcamRef,
   layout,
