+++ src/components/pages/SelectPage/DraggableImage.tsx
