+++ src/components/pages/SelectPage/StripLayout/DroppableImage.tsx
