+++ src/components/pages/SelectPage/StripLayout/StripLayout.tsx
